+++ src/app.ts
@@ -1,7 +1,8 @@
-import { doStackExchangeRequests } from './requests/stackExchange'
-import { dbConnect } from './config/databaseConfig'
 import dotenv from 'dotenv'
 import crypto from 'crypto'
+import { dbConnect } from './config/databaseConfig'
+import { doStackExchangeRequests } from './requests/stackExchange/stackExchange'
+import { doTwitterRequests } from './requests/twitter/index'
 
 dotenv.config()
 
@@ -13,4 +14,9 @@ console.log('Session ID is:', SESSION_ID)
 console.log('Use Control-C to exit')
 
 dbConnect()
+
+// TODO overwrite file w. array,
+// add to that array in storeMeta
+
 doStackExchangeRequests()
+// doTwitterRequests()
