+++ src/requests/twitter/index.ts
@@ -1,0 +1,40 @@
+import Twit from 'twit'
+import endpoints from './endpoints'
+import fetch, { Response } from 'node-fetch'
+import { storeResponseMeta } from '../../lib/storeMeta'
+
+interface Endpoint {
+  readonly method: string
+  readonly url: string
+  readonly params: object
+  readonly statusCode: number
+}
+
+export const doTwitterRequests = () => {
+  const twitterClient = new Twit({
+    consumer_key:
+      process.env.TWITTER_CONSUMER_API_KEY || '',
+    consumer_secret:
+      process.env.TWITTER_CONSUMER_API_SECRET || '',
+    access_token: process.env.TWITTER_ACCESS_TOKEN,
+    access_token_secret:
+      process.env.TWITTER_ACCESS_TOKEN_SECRET,
+  })
+
+  endpoints.forEach((endpoint: Endpoint) => {
+    try {
+      // TODO use node-fetch here instead, response not compatible
+
+      twitterClient.get(
+        endpoint.url,
+        endpoint.params,
+
+        function(err: any, data: any, response: any) {
+          // TODO call storeResponseMeta from here
+        }
+      )
+    } catch (e) {
+      console.log(e)
+    }
+  })
+}
